+++ docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Changes
+* Upgraded to [Module API Lib 0.10.0](https://github.com/kurtosis-tech/kurtosis-module-api-lib/blob/develop/docs/changelog.md#0100), which replaces all "Lambda" references with "module"
 
 # 0.1.7
 ### Fixes